GUI Design Brainstorm
This PowerPoint presentation offers initial insight into starting a draft of the GUI design to aid with a brainstorm for both the project requirements and the requirement specifications (specs). This offers a picture for components we will propose about including or not, and how it will influence the project requirements/specs. Not limited to the topics of architecture, software requirements, user platforms, speed, availability, response time, and incompatibility/compatibility, important element considerations outlined in project specifications. This included draft is revision number 1 (includes updates), where draft 0 is the initial proposal found in UMGC student locker of Group 4.